fix(server): move all handlers
From internal/server to internal/server/handlers
