+++ internal/app/app.go
@@ -7,7 +7,7 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/Employee-s-file-cabinet/backend/internal/config"
-	"github.com/Employee-s-file-cabinet/backend/internal/server"
+	"github.com/Employee-s-file-cabinet/backend/internal/server/handlers"
 	"github.com/Employee-s-file-cabinet/backend/internal/storage/db/postgresql"
 )
 
@@ -18,7 +18,7 @@ func Run(pctx context.Context, cfg *config.Config, logger *slog.Logger) error {
 	}
 	defer db.Close()
 
-	srv := server.New(cfg.HTTP, db, nil, logger)
+	srv := handlers.New(cfg.HTTP, db, nil, logger)
 
 	eg, ctx := errgroup.WithContext(pctx)
 	eg.Go(func() error {
