+++ internal/server/handlers/config.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 type Config struct {
 	Host  string `env:"HOST" env-default:"localhost"` // not used
