+++ internal/server/handlers/error.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 import (
 	"errors"
