+++ internal/server/handlers/health.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 import (
 	"net/http"
