+++ internal/server/handlers/middleware.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 import (
 	"fmt"
